% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugme-package.R
\docType{package}
\name{debugme-package}
\alias{debugme}
\alias{debugme-package}
\title{debugme: Debug R Packages}
\description{
Specify debug messages as special string constants, and control debugging of packages via environment variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/debugme#readme}
  \item \url{https://r-lib.github.io/debugme/}
  \item Report bugs at \url{https://github.com/r-lib/debugme/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

}
\keyword{internal}
