% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic.R
\name{handle_dynamic_code}
\alias{handle_dynamic_code}
\title{Handle the dynamic code in a debug string}
\usage{
handle_dynamic_code(str)
}
\arguments{
\item{str}{Debug string.}
}
\value{
A language expression or a string, depending on whether the
  string has dynamic code.
}
\description{
Substrings within backticks will be interpreted as code.
}
\keyword{internal}

