% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ debugme }
\alias{ debugme }
\alias{ debugme }
\title{ Debug R Packages }
\description{
Specify debug messages as special string constants, and control debugging of
  packages via environment variables.
}
