% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug}
\alias{debug}
\title{Debug message}
\usage{
debug(msg, pkg = environmentName(topenv(parent.frame())))
}
\arguments{
\item{msg}{Message to print, character constant.}

\item{pkg}{Package name to which the message belongs. Detected
automatically.}
}
\value{
The original message.
}
\description{
Normally this function is *not* called directly, but debug strings
are used. See [debugme()].
}

